requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
ultralytics>=8.2.0          # YOLOv8 plate detection
paddlepaddle==2.6.2         # PaddleOCR 2.7.x CPU backend (stable, no CUDA required)
paddleocr==2.7.3            # OCR engine (2.7.x API: use_gpu, show_log, use_angle_cls)

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python>=4.9.0
numpy>=1.24.0
Pillow>=10.0.0

# ── RAG / LLM (no local GPU needed – Groq free API) ───────────────────────────
groq>=0.9.0                 # Groq Cloud free tier  (Llama-3.1-8B / Mixtral-8x7b)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
chromadb>=0.5.0             # Local vector store
sentence-transformers>=3.0.0  # Embeddings  (paraphrase-multilingual-MiniLM – CPU OK)

# ── Backend ───────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
requests>=2.31.0

# ── Utils ─────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
loguru>=0.7.0
python-jose[cryptography]>=3.3.0
